ai-service/pyproject.toml
[project]
name = "chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "google-generativeai>=0.8.6",
    "pydantic-settings>=2.14.0",
    "uvicorn>=0.46.0",
]
